# src/main/resources/static/api-docs.yaml
openapi: 3.0.0
info:
  title: Demo Image Processing API
  version: 0.0.1
servers:
  - url: http://localhost:8080/api/v1
tags:
  - name: Image Controller
    description: Базовый CRUD API для работы с картинками
paths:
  /login:
    post:
      description: New endpoint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        200:
          description: Успешная авторизация
        404:
          description: Неверный логин
        401:
          description: Неверный пароль
  /image:
    post:
      tags:
        - Image Controller
      summary: Загрузка нового изображения в систему
      description: |
        В рамках данного метода необходимо:
        1. Провалидировать файл. Максимальный размер файла - 10Мб, поддерживаемые расширения - png, jpeg.
        1. Загрузить файл в S3 хранилище.
        1. Сохранить в БД мета-данные файла - название; размер; ИД файла в S3; ИД пользователя, которому файл принадлежит.
      operationId: uploadImage
      requestBody:
        content:
          multipart/form-data:
            schema:
              required:
                - "file"
              type: "object"
              properties:
                file:
                  type: "string"
                  format: "binary"
      responses:
        '200':
          description: Успех выполнения операции
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadImageResponse'
        '400':
          description: Файл не прошел валидацию
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UiSuccessContainer'
        '500':
          description: Непредвиденная ошибка
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UiSuccessContainer'
  /image/{image-id}:
    get:
      tags:
        - Image Controller
      summary: Скачивание файла по ИД
      description: |
        В рамках данного метода необходимо:
        1. Проверить, есть ли такой файл в системе.
        1. Проверить, доступен ли данный файл пользователю.
        1. Скачать файл.
      operationId: downloadImage
      parameters:
        - name: image-id
          in: path
          required: true
          schema:
            type: string
            format: id
      responses:
        '200':
          description: Успех выполнения операции
          content:
            '*/*':
              schema:
                type: string
                format: binary
        '404':
          description: Файл не найден в системе или недоступен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UiSuccessContainer'
        '500':
          description: Непредвиденная ошибка
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UiSuccessContainer'
    delete:
      tags:
        - Image Controller
      summary: Удаление файла по ИД
      description: |
        В рамках данного метода необходимо:
        1. Проверить, есть ли такой файл в системе.
        1. Проверить, доступен ли данный файл пользователю.
        1. Удалить файл.
      operationId: deleteImage
      parameters:
        - name: image-id
          in: path
          required: true
          schema:
            type: string
            format: id
      responses:
        '200':
          description: Успех выполнения операции
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UiSuccessContainer'
        '404':
          description: Файл не найден в системе или недоступен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UiSuccessContainer'
        '500':
          description: Непредвиденная ошибка
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UiSuccessContainer'
  /images:
    get:
      tags:
        - Image Controller
      summary: Получение списка изображений, которые доступны пользователю
      description: |
        В рамках данного метода необходимо:
        1. Получить мета-информацию о всех изображениях, которые доступны пользователю
      operationId: getImages
      responses:
        '200':
          description: Успех выполнения операции
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetImagesResponse'
        '500':
          description: Непредвиденная ошибка
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UiSuccessContainer'
components:
  schemas:
    LoginRequest:
      required:
        - username
        - password
      type: object
      properties:
        username:
          type: string
          description: Имя пользователя
        password:
          type: string
          description: Пароль
    UiSuccessContainer:
      required:
        - success
      type: object
      properties:
        success:
          type: boolean
          description: Признак успеха
        message:
          type: string
          description: Сообщение об ошибке
    UploadImageResponse:
      required:
        - imageId
      type: object
      properties:
        imageId:
          type: string
          format: id
          description: ИД файла
    GetImagesResponse:
      required:
        - images
      type: object
      properties:
        images:
          type: array
          description: Список изображений
          items:
            $ref: '#/components/schemas/Image'
    Image:
      required:
        - filename
        - size
      type: object
      properties:
        imageId:
          type: string
          format: id
          description: ИД файла
        filename:
          type: string
          description: Название изображения
        size:
          type: integer
          format: int32
          description: Размер файла в байтах